Refactor function to check conn status into a composable
