+++ kolibri/plugins/learn/assets/src/views/LearningActivityBar.vue
@@ -155,9 +155,8 @@
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
   import TimeDuration from 'kolibri.coreVue.components.TimeDuration';
   import SuggestedTime from 'kolibri.coreVue.components.SuggestedTime';
-  import { useConnectionChecker, useDevicesWithFacility } from 'kolibri.coreVue.componentSets.sync';
-  import { ref, watch } from 'kolibri.lib.vueCompositionApi';
   import get from 'lodash/get';
+  import useDeviceConnectionStatus from '../composables/useDeviceConnectionStatus';
   import LearningActivityIcon from './LearningActivityIcon.vue';
   import commonLearnStrings from './commonLearnStrings';
 
@@ -175,21 +174,9 @@
       SuggestedTime,
     },
     mixins: [KResponsiveWindowMixin, commonLearnStrings, commonCoreStrings],
-    setup(props, context) {
-      const disconnected = ref(false);
-      const { devices } = useDevicesWithFacility();
-      const { doCheck } = useConnectionChecker(devices);
+    setup(_, context) {
       const deviceId = get(context.root.$route, 'params.deviceId');
-      watch(devices, currentValue => {
-        if (!deviceId) return;
-        if (currentValue.length > 0) {
-          doCheck(deviceId).then(device => {
-            disconnected.value = !device.available;
-          });
-        } else {
-          disconnected.value = true;
-        }
-      });
+      const { disconnected } = useDeviceConnectionStatus(deviceId);
       return {
         disconnected,
       };
